Youtube modules plus small irc fix
